... By advancing time automatically when there's no work left. Changes computed experimentally, this is only the low-hanging fruit.

Here are the runtime gains (in s):
```
2.127 cluster::cluster_tests::basic_cluster_setup PASS [ 6.092s] narwhal-test-utils PASS [ 3.965s] narwhal-test-utils
2.173 test_get_collections_with_missing_certificates PASS [ 3.526s] narwhal-primary::integration_tests_validator_api PASS [ 1.353s] narwhal-primary::integration_tests_validator_api
2.292 test_simple_epoch_change PASS [ 12.098s] narwhal-primary::epoch_change PASS [ 9.806s] narwhal-primary::epoch_change
3.436 gateway_state::gateway_state_tests::test_public_transfer_object PASS [ 10.637s] sui-core PASS [ 7.201s] sui-core
3.494 epoch::reconfiguration_tests::test_start_epoch_change PASS [ 11.515s] sui-core PASS [ 8.021s] sui-core
5.059 test_shutdown_bug PASS [ 20.459s] narwhal-primary::nodes_bootstrapping_tests PASS [ 15.400s] narwhal-primary::nodes_bootstrapping_tests
7.076 gateway_state::gateway_state_tests::test_batch_transaction PASS [ 11.704s] sui-core PASS [ 4.628s] sui-core
10.001 mutex_table::test_mutex_table_bg_cleanup PASS [ 10.025s] sui-storage PASS [ 0.024s] sui-storage
10.064 gateway_state::gateway_state_tests::test_coin_split PASS [ 16.025s] sui-core PASS [ 5.961s] sui-core
```

Full list of expected gains: https://gist.github.com/745ad2177a62ed965bb29c234ea9edb7