@huitseeker 2022-11-10
> Does this reduce the chance to catch concurrency / timing issues in tests?

Only if you have tests which behavior depends on something occurring over a certain wall clock time, but that something is not work tracked by the async runtime (e.g. not in a task or awaited on). One example would be things depending on rocksdb background work (e.g. connecting a secondary to a primary, writing from the primary, expecting to read the update on the secondary).